Move ConfirVotes to its own folder
... and rename it to voteDialog to be alphabetically close to voting folder
