+++ src/components/voteDialog/index.js
