+++ src/components/voteDialog/voteAutocomplete.test.js
